--- src/test/java/com/grow/payment_service/PaymentApplicationServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.grow.payment_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PaymentApplicationServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
